chore: move sources to "src"
