+++ packages/cdk8s-plus/.npmignore
@@ -4,11 +4,16 @@
 # make sure to commit projen definition
 # standard-version configuration
 /.versionrc.json
+/src
+/test
+/coverage
+/.eslintrc.json
 # exclude typescript sources and configuration
 *.ts
 tsconfig.json
 # exclude jsii outdir
 dist
+!/lib
 !# include javascript files and typescript declarations
 !*.js
 !*.d.ts
