+++ packages/cdk8s-plus/package.json
@@ -11,14 +11,14 @@
   "scripts": {
     "projen": "node .projenrc.js && yarn install",
     "projen:upgrade": "yarn upgrade projen && yarn projen",
-    "test": "yarn eslint && jest --passWithNoTests",
+    "test": "yarn eslint && rm -fr lib/ && jest --passWithNoTests",
     "bump": "standard-version",
     "release": "yarn bump && git push --follow-tags origin master",
     "compile": "jsii --silence-warnings=reserved-word && jsii-docgen",
     "watch": "jsii -w --silence-warnings=reserved-word",
     "compat": "npx jsii-diff npm:$(node -p \"require('./package.json').name\")",
     "package": "jsii-pacmak",
-    "build": "jsii && yarn docgen",
+    "build": "jsii --silence-warnings=reserved-word && yarn docgen",
     "eslint": "eslint . --ext .ts",
     "docgen": "jsii-docgen",
     "test:watch": "jest --watch",
@@ -32,7 +32,7 @@
     "constructs": "2.0.1",
     "cdk8s": "0.0.0",
     "@types/minimatch": "^3.0.3",
-    "projen": "^0.1.36",
+    "projen": "^0.2.2",
     "standard-version": "^8.0.0",
     "jsii": "^1.6.0",
     "jsii-diff": "^1.6.0",
@@ -88,6 +88,10 @@
         "namespace": "Org.Cdk8s.Plus",
         "packageId": "Org.Cdk8s.Plus"
       }
+    },
+    "tsc": {
+      "outDir": "lib",
+      "rootDir": "src"
     }
   },
   "jest": {
@@ -103,6 +107,11 @@
     "testMatch": [
       "**/__tests__/**/*.ts?(x)",
       "**/?(*.)+(spec|test).ts?(x)"
-    ]
+    ],
+    "globals": {
+      "ts-jest": {
+        "tsConfig": "tsconfig.jest.json"
+      }
+    }
   }
 }
